+++ packages/react/src/withStyles.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import hoistNonReactStatics from 'hoist-non-react-statics';
 import { LocalSheet, StringOnly } from '@aesthetic/core';
 import useStyles from './useStyles';
-import { WithStylesWrappedProps, WithStylesWrapperProps, WrapperComponent } from './types';
+import { WithStylesWrappedProps, WrapperProps, WrapperComponent } from './types';
 
 /**
  * Wrap a React component with an HOC that injects the style to class name transfer function.
@@ -12,14 +12,14 @@ export default function withStyles<T = unknown>(factory: LocalSheet<T>) /* infer
 
   return function withStylesComposer<Props extends object = {}>(
     WrappedComponent: React.ComponentType<Props & WithStylesWrappedProps<StringKeys>>,
-  ): React.NamedExoticComponent<Props & WithStylesWrapperProps> & WrapperComponent {
+  ): React.NamedExoticComponent<Props & WrapperProps> & WrapperComponent {
     const WithStyles = React.memo(function WithStyles({
       wrappedRef,
       ...props
-    }: Props & WithStylesWrapperProps) {
+    }: Props & WrapperProps) {
       const cx = useStyles(factory);
 
-      return <WrappedComponent {...(props as any)} cx={cx} ref={wrappedRef} />;
+      return <WrappedComponent {...(props as any)} ref={wrappedRef} cx={cx} />;
     }) as ReturnType<typeof withStylesComposer>;
 
     hoistNonReactStatics(WithStyles, WrappedComponent);