+++ packages/react/src/withTheme.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import hoistNonReactStatics from 'hoist-non-react-statics';
 import useTheme from './useTheme';
-import { WithThemeWrappedProps, WithThemeWrapperProps, WrapperComponent } from './types';
+import { WithThemeWrappedProps, WrapperProps, WrapperComponent } from './types';
 
 /**
  * Wrap a React component with an HOC that injects the current theme object as a prop.
@@ -9,8 +9,8 @@ import { WithThemeWrappedProps, WithThemeWrapperProps, WrapperComponent } from '
 export default function withTheme() /* infer */ {
   return function withThemeComposer<Props extends object = {}>(
     WrappedComponent: React.ComponentType<Props & WithThemeWrappedProps>,
-  ): React.FunctionComponent<Props & WithThemeWrapperProps> & WrapperComponent {
-    function WithTheme({ wrappedRef, ...props }: Props & WithThemeWrapperProps) {
+  ): React.FunctionComponent<Props & WrapperProps> & WrapperComponent {
+    function WithTheme({ wrappedRef, ...props }: Props & WrapperProps) {
       const theme = useTheme();
 
       return <WrappedComponent {...(props as any)} ref={wrappedRef} theme={theme} />;